requirements.txt
pillow
numpy
wand
